requirements.txt
python-docx
pyTelegramBotAPI
lxml
google-generativeai
python-dotenv
flask
flask-cors
